refactor(table): rename scss files to use .module.scss
