+++ draft-packages/table/KaizenDraft/Table/Table.tsx
@@ -1,13 +1,13 @@
+import React from "react"
 import { Icon } from "@kaizen/component-library"
 import { Heading } from "@kaizen/typography"
 import { Checkbox, CheckedStatus } from "@kaizen/draft-form"
 import classNames from "classnames"
-import * as React from "react"
 import sortAscendingIcon from "@kaizen/component-library/icons/sort-ascending.icon.svg"
 import sortDescendingIcon from "@kaizen/component-library/icons/sort-descending.icon.svg"
 import exclamationIcon from "@kaizen/component-library/icons/exclamation.icon.svg"
 import { Tooltip } from "@kaizen/draft-tooltip"
-import styles from "./styles.scss"
+import styles from "./Table.module.scss"
 
 type TableContainer = React.FunctionComponent<TableContainerProps>
 type TableContainerProps = {
